# [murmur] Fix MurmurBytes