--- a/murmur_test.go
+++ b/murmur_test.go
@@ -69,7 +69,7 @@ func TestMurmurString(t *testing.T) {
 	s := "hello"
 	v := MurmurString(s)
 	if v != 613153351 {
-		t.Fatalf("MurmurBytes failed for %s: %v != %v", s, v, 613153351)
+		t.Fatalf("MurmurString failed for %s: %v != %v", s, v, 613153351)
 	}
 }
 
@@ -83,7 +83,7 @@ func randString(n int) string {
 }
 
 // Benchmarks
-func benchmarkMurmer64(b *testing.B, input []uint64) {
+func benchmarkMurmur64(b *testing.B, input []uint64) {
 	b.ResetTimer()
 	for n := 0; n < b.N; n++ {
 		for _, x := range input {
@@ -92,7 +92,7 @@ func benchmarkMurmer64(b *testing.B, input []uint64) {
 	}
 }
 
-func benchmarkMurmerString(b *testing.B, input []string) {
+func benchmarkMurmurString(b *testing.B, input []string) {
 	b.ResetTimer()
 	for n := 0; n < b.N; n++ {
 		for _, x := range input {
@@ -111,20 +111,20 @@ func benchmarkHash32(b *testing.B, input []string) {
 	}
 }
 
-func Benchmark100Murmer64(b *testing.B) {
+func Benchmark100Murmur64(b *testing.B) {
 	input := make([]uint64, 100)
 	for i := 0; i < 100; i++ {
 		input[i] = uint64(rand.Int63())
 	}
-	benchmarkMurmer64(b, input)
+	benchmarkMurmur64(b, input)
 }
 
-func Benchmark100MurmerString(b *testing.B) {
+func Benchmark100MurmurString(b *testing.B) {
 	input := make([]string, 100)
 	for i := 0; i < 100; i++ {
 		input[i] = randString((i % 15) + 5)
 	}
-	benchmarkMurmerString(b, input)
+	benchmarkMurmurString(b, input)
 }
 
 func Benchmark100Hash32(b *testing.B) {