@Viq111 2019-09-03
Latest benchmarks:

Previous vs this PR:
```go
ᐅ benchstat old new
name               old time/op  new time/op  delta
100MurmurString-4  1.08µs ± 5%  1.26µs ± 4%  +17.24%  (p=0.000 n=10+10)
```

Current PR vs first iteration:
```go
ᐅ benchstat new2 new1
name               old time/op  new time/op  delta
100MurmurString-4  1.37µs ± 2%  1.26µs ± 4%  -7.86%  (p=0.000 n=10+10)
```

So for correctness we pay 17% perf decrease (200ns), and the latest iteration is the fastest of all 3 others:
```go
Kenny:
    sh := (*reflect.StringHeader)(unsafe.Pointer(&key))
    byteSlice := (*[math.MaxInt32]byte)(unsafe.Pointer(sh.Data))[:sh.Len:sh.Len]

New:
    b := make([]byte, 0)
    bHeader := (*reflect.SliceHeader)(unsafe.Pointer(&b))
    sHeader := (*reflect.StringHeader)(unsafe.Pointer(&key))
    bHeader.Data = sHeader.Data
    bHeader.Len = sHeader.Len
    bHeader.Cap = sHeader.Len
    return MurmurBytes(b)

Copy:
    return MurmurBytes([]byte(key))
```